[app] indexer: move insertData to save.go and add time track
