+++ internal/app/indexer/fetch.go
@@ -186,56 +186,3 @@ func (s *Service) fetchMasterLoop(fromBlock uint32, results chan<- processedMast
 		fromBlock = blocks[len(blocks)-1].master.block.SeqNo + 1
 	}
 }
-
-func (s *Service) insertData(
-	acc []*core.AccountState,
-	msg []*core.Message,
-	tx []*core.Transaction,
-	b []*core.Block,
-) error {
-	ctx := context.Background()
-
-	dbTx, err := s.DB.PG.Begin()
-	if err != nil {
-		return errors.Wrap(err, "cannot begin db tx")
-	}
-	defer func() {
-		_ = dbTx.Rollback()
-	}()
-
-	for _, message := range msg {
-		err := s.Parser.ParseMessagePayload(ctx, message)
-		if errors.Is(err, app.ErrImpossibleParsing) {
-			continue
-		}
-		if err != nil {
-			log.Error().Err(err).
-				Hex("msg_hash", message.Hash).
-				Hex("src_tx_hash", message.SrcTxHash).
-				Str("src_addr", message.SrcAddress.String()).
-				Hex("dst_tx_hash", message.DstTxHash).
-				Str("dst_addr", message.DstAddress.String()).
-				Uint32("op_id", message.OperationID).
-				Msg("parse message payload")
-		}
-	}
-
-	if err := s.accountRepo.AddAccountStates(ctx, dbTx, acc); err != nil {
-		return errors.Wrap(err, "add account states")
-	}
-	if err := s.msgRepo.AddMessages(ctx, dbTx, msg); err != nil {
-		return errors.Wrap(err, "add messages")
-	}
-	if err := s.txRepo.AddTransactions(ctx, dbTx, tx); err != nil {
-		return errors.Wrap(err, "add transactions")
-	}
-	if err := s.blockRepo.AddBlocks(ctx, dbTx, b); err != nil {
-		return errors.Wrap(err, "add shard block")
-	}
-
-	if err := dbTx.Commit(); err != nil {
-		return errors.Wrap(err, "cannot commit db tx")
-	}
-
-	return nil
-}
